+++ src/services/form-validation/schema.ts
@@ -1,294 +1,22 @@
-import validate, { ValidatorConstraintFn } from 'validate.js'
+import { ValidatorConstraintFn } from 'validate.js'
 import { attachmentsService, localisationService } from '@oneblink/apps'
-import { FormTypes, FormsAppsTypes, SubmissionTypes } from '@oneblink/types'
-import { FormElementBinaryStorageValue } from '../types/attachments'
-import { Value as FormElementComplianceValue } from '../form-elements/FormElementCompliance'
-import { parseDateValue } from './generate-default-data'
-import generateCivicaNameRecordElements from './generateCivicaNameRecordElements'
+import { FormTypes, SubmissionTypes } from '@oneblink/types'
+import { parseDateValue } from '../generate-default-data'
+import generateCivicaNameRecordElements from '../generateCivicaNameRecordElements'
+
+import generateFreshdeskDependentFieldElements from '../generateFreshdeskDependentFieldElements'
 import {
-  CaptchaType,
-  ExecutedLookups,
-  ExecutedLookupValue,
-  FormElementConditionallyShown,
-  FormElementsConditionallyShown,
-  FormElementsValidation,
-} from '../types/form'
-import generateFreshdeskDependentFieldElements from './generateFreshdeskDependentFieldElements'
-import cleanFormSubmissionModel from './cleanFormSubmissionModel'
-import getDateRangeConfiguration, {
-  DateRangeConfigurationOptions,
-} from './getDateRangeConfiguration'
-import getRepeatableSetEntriesConfiguration from './getRepeatableSetEntriesConfiguration'
+  presence,
+  escapeElementName,
+  getCleanDateRangeConfiguration,
+  getCleanRepeatableSetConfiguration,
+  getCustomRegexFormatConfig,
+  checkFileNameIsValid,
+} from './validators'
 
-type NestedValidateJSSchema = {
-  schema: ValidateJSSchema
-  formElementConditionallyShown: FormElementConditionallyShown | undefined
-  executedLookups: ExecutedLookupValue
-  captchaType: CaptchaType
-}
+export type ValidateJSSchema = Record<string, unknown>
 
-export const generateLookupValidationMessage = (
-  lookupButtonConfig?: FormsAppsTypes.ButtonConfiguration,
-) => {
-  return lookupButtonConfig && lookupButtonConfig.label
-    ? `${lookupButtonConfig.label} is required`
-    : 'Lookup is required'
-}
-
-// https://validatejs.org/#validators-datetime
-// Before using it we must add the parse and format functions
-// Here is a sample implementation using moment.js
-validate.extend(validate.validators.datetime, {
-  // The value is guaranteed not to be null or undefined but otherwise it
-  // could be anything.
-  parse: function (value: string) {
-    return Date.parse(value)
-  },
-  // Input is a unix timestamp
-  format: function (
-    value: string,
-    options: { format: (date: Date) => string },
-  ) {
-    const dateValue = new Date(value)
-    return options.format(dateValue)
-  },
-})
-
-validate.validators.entries = function (
-  value: unknown,
-  {
-    setSchema,
-    entrySchema: {
-      schema: entrySchema,
-      formElementConditionallyShown,
-      executedLookups,
-      captchaType,
-    },
-  }: {
-    setSchema: ValidateJSSchema
-    entrySchema: NestedValidateJSSchema
-  },
-) {
-  const entries = Array.isArray(value) ? value : []
-
-  const entryErrors = entries.reduce((errorsByIndex, entry, index) => {
-    const entryValidation = validateSubmission(
-      entrySchema,
-      entry,
-      formElementConditionallyShown?.type === 'repeatableSet'
-        ? formElementConditionallyShown.entries[index.toString()]
-        : undefined,
-      Array.isArray(executedLookups) ? executedLookups[index] : {},
-      captchaType,
-    )
-    if (entryValidation) {
-      errorsByIndex[index] = entryValidation
-    }
-    return errorsByIndex
-  }, {})
-
-  const setErrorMessages = validate.single(value, setSchema)
-
-  if (!setErrorMessages && validate.isEmpty(entryErrors)) {
-    return
-  }
-
-  return {
-    type: 'repeatableSet',
-    set: setErrorMessages && setErrorMessages[0],
-    entries: entryErrors,
-  }
-}
-
-validate.validators.nestedElements = function (
-  value: SubmissionTypes.S3SubmissionData['submission'] | undefined,
-  {
-    schema,
-    formElementConditionallyShown,
-    executedLookups,
-    captchaType,
-  }: NestedValidateJSSchema,
-) {
-  const errors = validateSubmission(
-    schema,
-    value,
-    formElementConditionallyShown?.type === 'formElements'
-      ? formElementConditionallyShown.formElements
-      : undefined,
-    typeof executedLookups !== 'boolean' ? executedLookups : {},
-    captchaType,
-  )
-  if (!errors) {
-    return
-  }
-  return {
-    type: 'formElements',
-    formElements: errors,
-  }
-}
-
-function getInvalidAttachment(value: FormElementBinaryStorageValue) {
-  if (
-    value &&
-    typeof value === 'object' &&
-    value.type &&
-    value.type === 'ERROR'
-  ) {
-    return value
-  }
-}
-function validateAttachments(
-  value: FormElementBinaryStorageValue[] | undefined,
-) {
-  const invalidAttachmentNames = value?.reduce(
-    (invalidAttachmentNames: string[], att) => {
-      const attachmentName = getInvalidAttachment(att)?.fileName
-      if (attachmentName) {
-        invalidAttachmentNames.push(attachmentName)
-      }
-      return invalidAttachmentNames
-    },
-    [],
-  )
-  if (invalidAttachmentNames?.length) {
-    return `${invalidAttachmentNames.join(', ')} could not be uploaded.`
-  }
-}
-validate.validators.attachment = function (
-  value: FormElementBinaryStorageValue,
-) {
-  return getInvalidAttachment(value)?.errorMessage
-}
-validate.validators.attachments = function (
-  value:
-    | FormElementBinaryStorageValue[]
-    | FormElementComplianceValue
-    | undefined,
-) {
-  if (Array.isArray(value)) {
-    return validateAttachments(value)
-  }
-  return validateAttachments(
-    value?.files as FormElementBinaryStorageValue[] | undefined,
-  )
-}
-
-// Extend validator for lookups
-validate.validators.lookups = function (
-  value: unknown,
-  {
-    executedLookups,
-    formElement,
-  }: {
-    formElement: FormTypes.LookupFormElement & FormTypes.FormElementRequired
-    executedLookups: ExecutedLookups
-  },
-) {
-  if (!formElement.isDataLookup && !formElement.isElementLookup) {
-    return
-  }
-
-  // Lookups must only be executed on required form elements
-  if (formElement && !formElement.required) {
-    return
-  }
-
-  const elementExecutedLookups = executedLookups?.[formElement.name]
-  if (elementExecutedLookups === true) {
-    return
-  }
-  return generateLookupValidationMessage(formElement.lookupButton)
-}
-
-validate.validators.numberRegex = function (value: unknown, format: unknown) {
-  if (!format) {
-    return
-  }
-  if (typeof value === 'number' && !Number.isNaN(value)) {
-    value = value.toString()
-  }
-  const errorMessages = validate.single(value, { format })
-  return errorMessages && errorMessages[0]
-}
-
-validate.validators.isTrue = function (value: unknown, message?: string) {
-  if (!value) {
-    return message || 'Must be set to true'
-  }
-}
-
-validate.validators.needsExtension = function (
-  value: attachmentsService.Attachment[] | undefined,
-  formElement: FormTypes.FilesElement,
-) {
-  const isValid =
-    !Array.isArray(value) ||
-    value.every((file) => {
-      return checkFileNameExtensionIsValid(formElement, file.fileName)
-    })
-  if (!isValid) return 'All files must have extensions'
-}
-
-function getCustomRegexFormatConfig<DefaultValue>(
-  formElement: FormTypes.FormElementWithInput<DefaultValue>,
-) {
-  return formElement.regexPattern
-    ? {
-        pattern: formElement.regexPattern,
-        flags: formElement.regexFlags,
-        message: formElement.regexMessage,
-      }
-    : undefined
-}
-
-type ValidateJSSchema = Record<string, unknown>
-
-const presence = (
-  { required, requiredMessage }: FormTypes.FormElementRequired,
-  message: string,
-) => (required ? { message: requiredMessage || message } : false)
-
-const escapeElementName = (elementName: string) => {
-  const escapedName = elementName.replace(/\./g, '\\.')
-  return escapedName
-}
-
-function getCleanDateRangeConfiguration(
-  options: DateRangeConfigurationOptions,
-  elements: FormTypes.FormElement[],
-  submission: SubmissionTypes.S3SubmissionData['submission'] | undefined,
-  formElementsConditionallyShown: FormElementsConditionallyShown | undefined,
-): ReturnType<typeof getDateRangeConfiguration> {
-  if (options.referenceFormElementId && submission) {
-    const { model } = cleanFormSubmissionModel(
-      submission,
-      elements,
-      formElementsConditionallyShown,
-      true,
-    )
-    return getDateRangeConfiguration(options, elements, model)
-  }
-  return [options.date, options.daysOffset]
-}
-
-function getCleanRepeatableSetConfiguration(
-  setEntries: FormTypes.RepeatableSetElement['minSetEntries'],
-  elements: FormTypes.FormElement[],
-  submission: SubmissionTypes.S3SubmissionData['submission'] | undefined,
-  formElementsConditionallyShown: FormElementsConditionallyShown | undefined,
-) {
-  if (submission) {
-    const { model } = cleanFormSubmissionModel(
-      submission,
-      elements,
-      formElementsConditionallyShown,
-      true,
-    )
-    return getRepeatableSetEntriesConfiguration(setEntries, elements, model)
-  }
-}
-
-export function generateValidationSchema(
+export default function generateValidationSchema(
   elements: FormTypes.FormElementWithName[],
 ): ValidateJSSchema {
   return elements.reduce<ValidateJSSchema>((partialSchema, formElement) => {
@@ -776,84 +504,4 @@ export function generateValidationSchema(
     partialSchema[escapeElementName(formElement.name)] = constraint
     return partialSchema
   }, {})
-}
-
-export function validateSubmission(
-  schema: ValidateJSSchema,
-  submission: SubmissionTypes.S3SubmissionData['submission'] | undefined,
-  formElementsConditionallyShown: FormElementsConditionallyShown | undefined,
-  executedLookups: ExecutedLookupValue,
-  captchaType: CaptchaType,
-): FormElementsValidation | undefined {
-  const errorsAsArray = validate(submission, schema, {
-    format: 'grouped',
-    fullMessages: false,
-    formElementsConditionallyShown,
-    executedLookups,
-    captchaType,
-  })
-  if (!errorsAsArray || validate.isEmpty(errorsAsArray)) {
-    return
-  }
-
-  const errors = Object.keys(errorsAsArray || {}).reduce(
-    (messagesByFormElementName: FormElementsValidation, key) => {
-      const validationMessage = errorsAsArray[key][0]
-      if (validationMessage) {
-        messagesByFormElementName[key.replace(/\\./g, '.')] = validationMessage
-      }
-      return messagesByFormElementName
-    },
-    {},
-  )
-  if (validate.isEmpty(errors)) {
-    return
-  }
-
-  return errors
-}
-
-export function checkFileNameIsValid(
-  formElement: FormTypes.FilesElement,
-  fileName: string,
-) {
-  const extension = fileName.split('.').pop()
-  return (
-    !formElement.restrictedFileTypes ||
-    formElement.restrictedFileTypes.some(
-      (fileType) => fileType.toLowerCase() === extension?.toLowerCase(),
-    )
-  )
-}
-
-export function checkFileNameExtensionIsValid(
-  formElement: FormTypes.FilesElement,
-  fileName: string,
-) {
-  return (
-    formElement.allowExtensionlessAttachments || fileName.split('.').length >= 2
-  )
-}
-
-export function checkSectionValidity(
-  element: FormTypes.PageElement | FormTypes.SectionElement,
-  formElementsValidation: FormElementsValidation | undefined,
-): boolean {
-  // If everything is valid, no need to check elements
-  if (!formElementsValidation) {
-    return false
-  }
-
-  // If there is no elements on the page that are invalid, we will treat as valid
-  return element.elements.some((formElement) => {
-    switch (formElement.type) {
-      case 'page':
-      case 'section': {
-        return checkSectionValidity(formElement, formElementsValidation)
-      }
-      default: {
-        return formElementsValidation[formElement.name]
-      }
-    }
-  })
 }
